pyproject.toml
[project]
name = "birthday-present-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.10"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.9.0",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.0",
    "pydantic>=2.10.6",
    "tenacity>=9.0.0",
]
